Muoa Version 0.3.1 - Wrote the first version of Muoa's language specification - Muoa's number type is now a 64-but floating point instead of a 32-bit floating point

+++ Runtime/Interpreter.cs
@@ -18,7 +18,7 @@ public class Interpreter(Scope? scope, Dictionary<Lexer.Token.Type, IMuoaFunctio
                 switch (node)
                 {
                     case NumberNode num:
-                        Scope.Push(new MuoaNumber(float.Parse(num.value.lit)));
+                        Scope.Push(new MuoaNumber(double.Parse(num.value.lit)));
                         break;
                     case AtomNode atom:
                         Scope.Push(new MuoaAtom(atom.value.lit));
